changed footer and its position
